+++ cc.inspoy.framework.ui/Tweener/Actions/Delay.cs
@@ -1,0 +1,17 @@
+/**
+ * == Inspoy Technology ==
+ * Assembly: Instech.Framework.Ui
+ * FileName: Delay.cs
+ * Created on 2020/10/22 by inspoy
+ * All rights reserved.
+ */
+
+namespace Instech.Framework.Ui.Tweening
+{
+    public class Delay : PersistentAction
+    {
+        public Delay(float interval) : base(interval)
+        {
+        }
+    }
+}